store interface and implementation in the same file
