+++ VideoTag.Server/Repositories/CategoryRepository.cs
@@ -4,6 +4,15 @@ using VideoTag.Server.Entities;
 
 namespace VideoTag.Server.Repositories;
 
+public interface ICategoryRepository
+{
+    Task InsertCategory(Category category);
+    Task<IEnumerable<Category>> GetCategories(bool includeTags = false);
+    Task<Category> GetCategory(Guid categoryId);
+    Task UpdateCategory(Category category);
+    Task DeleteCategory(Guid categoryId);
+}
+
 public class CategoryRepository(DapperContext dapperContext) : ICategoryRepository
 {
     public async Task InsertCategory(Category category)
