--- VideoTag.Server/Repositories/ICategoryRepository.cs
@@ -1,16 +1,0 @@
-using VideoTag.Server.Entities;
-
-namespace VideoTag.Server.Repositories;
-
-public interface ICategoryRepository
-{
-    Task InsertCategory(Category category);
-
-    Task<IEnumerable<Category>> GetCategories(bool includeTags = false);
-
-    Task<Category> GetCategory(Guid categoryId);
-
-    Task UpdateCategory(Category category);
-    
-    Task DeleteCategory(Guid categoryId);
-}
