+++ VideoTag.Server/Repositories/TagRepository.cs
@@ -4,6 +4,15 @@ using VideoTag.Server.Entities;
 
 namespace VideoTag.Server.Repositories;
 
+public interface ITagRepository
+{
+    Task InsertTag(Tag tag);
+    Task<IEnumerable<Tag>> GetTags();
+    Task<Tag> GetTag(Guid tagId);
+    Task UpdateTag(Tag tag);
+    Task DeleteTag(Guid tagId);
+}
+
 public class TagRepository(DapperContext dapperContext) : ITagRepository
 {
     public async Task InsertTag(Tag tag)
